+++ actors/autoesa-daily/package.json
@@ -27,6 +27,6 @@
 		"build": "tsc",
 		"test": "echo \"Error: oops, the actor has no tests yet, sad!\" && exit 1"
 	},
-	"author": "It's not you it's me",
+	"author": "Martin Forejt <martin.forejt@apify.com>",
 	"license": "ISC"
 }
